Frontend/requirements.txt
chess
flask
flask-cors
python-dotenv
requests